src/backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.27.0
redis==5.0.11

# Configuration
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Client
requests==2.31.0
httpx==0.26.0

# WebSocket
websockets==12.0

# Utilities
python-multipart==0.0.6
